requirements.txt
django
djangorestframework
python-dotenv
drf-spectacular
requests
redis
celery
django-celery-beat
flower
